created user service and user service impl show profile query + security set roles into users

+++ common/src/main/java/com/zemelya/repository/user/UserSpringDataRepository.java
@@ -39,5 +39,5 @@ public interface UserSpringDataRepository extends JpaRepository<HibernateUser, L
   @Query(
       value = "insert into rentalcars.l_role_user(user_id, role_id) values (:user_id, :role_id)",
       nativeQuery = true)
-  int createRoleRow(@Param("user_id") Long userId, @Param("role_id") Long roleId);
+  int createRoleRow(@Param("user_id") Long userId, @Param("role_id") Integer roleId);
 }
